+++ Classes/Models/CDXCard.m
@@ -123,7 +123,7 @@
             if (fontSizeCacheSize[i].height == size.height &&
                 fontSizeCacheSize[i].width == size.width) {
                 CGFloat cachedFontSize = fontSizeCacheFontSize[i];
-                qltrace("cached: %3.0f index: %d", cachedFontSize, i);
+                qltrace("cached: %3.0f index: %lu", cachedFontSize, (unsigned long)i);
                 return cachedFontSize;
             }
         }
@@ -168,7 +168,7 @@
         if (fontSizeCacheNextIndex < CDXCardFontSizeCacheSize) {
             fontSizeCacheFontSize[fontSizeCacheNextIndex] = minlineFontSize;
             fontSizeCacheSize[fontSizeCacheNextIndex] = size;
-            qltrace(@"calculated: %3.0f index: %d", minlineFontSize, fontSizeCacheNextIndex);
+            qltrace(@"calculated: %3.0f index: %lu", minlineFontSize, (unsigned long)fontSizeCacheNextIndex);
             fontSizeCacheNextIndex++;
         } else {
             qltrace(@"calculated: %3.0f", minlineFontSize);
